Change expected parameter values to match a buggy output in the export/import functionality.
The problem lies with the selection of dropdown values when multiple profiles have the same name as well as a problem with the default values in the GUI not matching up between the hidden dropdown value range and what the dropdown object has selected. This is a larger bug that cannot be fixed in this commit as it requires GUI changes. Once fixed, the expected values might have to be changed again.