Remove conversion from dot to comma as decimal seperator in import of schedules.
This originally intended to solve a bug with differences in what seperator Excel expects for numbers, however in VBA code the numbers are only detected as numbers if they have a dot as the seperator, thus making the conversion not only unnecessary but actually introducing a bug. This bug is now fixed.