Restructure measure set_weather_axis_timestep to implement three different sizing methods.
The first is the old implementation of using a DDY file.
The second uses automatic determination of design days from the weather data. This isn't working yet.
The third uses manually specified design days. This is somewhat worse than both options but has the advantage that it avoids the need for DDY files and it's working reasonably well.